fix: rename AccountsConnectedItemList to AccountsConnectedList
