--- app/components/Views/AccountConnect/AccountsConnectedItemList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AccountsConnectedItemList';
